requirements.txt
pandas
python-barcode[images]
python-docx
pillow
streamlit
